refactor: added module allocator
